Succesfully refactored. All rules tests are passing.
