+++ tests/rules/test_git_rm_local_modifications.py
@@ -1,5 +1,5 @@
 import pytest
-from thefuck.rules.git_rm_local_modifications import match, get_new_command
+from thefuck.rules.git_rm_staged_and_local_modifications import match, get_new_command
 from thefuck.types import Command
