+++ thefuck/rules/git_rm_staged_and_local_modifications.py
@@ -4,7 +4,8 @@ from thefuck.specific.git import git_support
 @git_support
 def match(command):
     return (' rm ' in command.script and
-            'error: the following file has changes staged in the index' in command.output and
+            ('error: the following file has changes staged in the index' in command.output or
+             'error: the following file has local modifications' in command.output) and
             'use --cached to keep the file, or -f to force removal' in command.output)
